Additional building class improvements (check description)
Changed how entities are passed around.

Before, PassEntity() was called when an entity was being passed from one building to another, and then ReceiveEntity() was called once that entity got to EITHER the input OR output.

Now, PassEntity() has been renamed to InputEntity(), and ReceiveEntity() has been split into ReceieveEntity() and OutputEntity().

To recap, InputEntity() is called on a building from another building that wants to give it an entity, ReceiveEntity() is called by the EntityHandler once that entity has arrived at an input position, and OutputEntity() is called when an entity arrives at an output position and is ready to be passed to another building.